feat) 레이저건 쏘는 방향에 대한 line rendering 추가
쏘는 방향으로 line이 나가며, 그 방향으로 raycastAll이 일어나게 했음
아직 데미지 여부는 모름

+++ Assets/Script/Interfaces.cs
@@ -2,10 +2,10 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(int amage);
+    void TakeDamage(float amage);
 }
 
 public interface IHealable
 {
-    void TakeHeal(int heal);
+    void TakeHeal(float heal);
 }
